* commit '547cfadda0f8cd4a83ef73c40267b26de80dc6a6':
  YUK-21418: Sep-Temperature-Offset-the-cooling-offset issue fix

+++ yukon-client/common/src/com/cannontech/common/dr/gear/setup/fields/ProgramGearFieldsBuilder.java
@@ -580,8 +580,8 @@ public class ProgramGearFieldsBuilder {
         boolean isBackRampEnabled = backRampOption.compareTo(IlmDefines.RAMP_RANDOM) == 0;
         gearFields.setRampOut(isBackRampEnabled);
         LMThermostatGear thermostatGear = gearDao.getLMThermostatGear(directGear.getGearId());
-        Integer valueTa = thermostatGear.getValueTa();
-        Integer valueTb = thermostatGear.getValueTb();
+        double valueTa = thermostatGear.getValueTa();
+        double valueTb = thermostatGear.getValueTb();
 
         double heatingOffset = valueTa / 10;
         double coolingOffset = valueTb / 10;
